+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a preview control to enable or disable dragging entire objects in edit mode, disabled by default.
+
 ## [0.8.2] - 2026-09-03
 
 ### Changed